docker1/requirements.txt
fastapi
uvicorn
kafka-python
requests
mysql-connector-python
